+++ app/login/page.tsx
@@ -1,10 +1,11 @@
 //Page Accueil/Login
-import LoginForm from '../auth/components/LoginForm'; 
+import LoginForm from '../../auth/components/LoginForm'; 
 
 // Création de la page de connexion qui utilise le composant loginForm
 const LoginPage = () => {
   return (
     <LoginForm />
+    
   );
 };
