+++ tailwind.config.ts
@@ -7,6 +7,9 @@ const config: Config = {
     "./components/**/*.{js,ts,jsx,tsx,mdx}",
     "./app/**/*.{js,ts,jsx,tsx,mdx}",
     "./calculator/**/*.{js,ts,jsx,tsx,mdx}",
+    "./auth/**/*.{js,ts,jsx,tsx,mdx}",
+    "./common/**/*.{js,ts,jsx,tsx,mdx}",
+    
   ],
   theme: {
     extend: {
